chore: use same ArticleCard across website
