+++ components/common/ArticleCard.tsx
@@ -1,91 +1,59 @@
 import Link from 'next/link'
 
-import Date from './Date'
-import ImageComponent from './ImageComponent'
-import { badgeVariants } from '@/components/ui/badge'
-import { cn } from '@/lib/utils'
+import { Separator } from '@/components/ui/separator'
+import { Badge } from '@/components/ui/badge'
+import ImageComponent from '@/components/common/ImageComponent'
+import Date from '@/components/common/Date'
 
-type ArticleCardProps = {
-  title: string
-  date: string
-  image: any
-  slug: string
-  division: {
-    name: string
-    slug: string
-  }
-  conferences: {
-    name: string
-    shortName: string
-    slug: string
-  }[]
-  author: {
-    name: string
-    slug: string
-    archived: boolean
-  }
-  index?: number
-}
-
-const ArticleCard = ({
+export default function ArticleCard({
   title,
-  date,
   image,
   slug,
+  author,
+  date,
   division,
   conferences,
-  author,
-  index,
-}: ArticleCardProps) => {
+}: {
+  title: string
+  image: any
+  slug: string
+  author: string
+  date: string
+  division: {
+    name: string
+  }
+  conferences: {
+    shortName: string
+  }[]
+}) {
   return (
-    <div className="rounded-xl border bg-card text-card-foreground shadow">
-      <Link
-        href={`/${slug}`}
-        className="aspect-h-1 aspect-w-2 relative block overflow-hidden rounded-t-xl shadow-md"
-      >
-        <ImageComponent
-          image={image}
-          alt={image.caption}
-          width={363}
-          height={181}
-          loading={index !== undefined && index < 3 ? 'eager' : 'lazy'}
-        />
-      </Link>
-      <div className="space-y-4 p-4">
-        <div className="flex flex-wrap gap-3">
-          {division && (
-            <Link
-              href={`/news/${division.slug}`}
-              className={cn(badgeVariants({ variant: 'default' }), 'font-semibold')}
-            >
-              {division.name}
-            </Link>
-          )}
+    <div className="overflow-hidden rounded-lg border border-border shadow-lg">
+      <ImageComponent
+        image={image}
+        alt={image.caption}
+        width={600}
+        height={400}
+        className="h-48 w-full object-cover object-top"
+      />
+      <div className="bg-background p-4">
+        <div className="mb-2 flex flex-wrap gap-2">
+          {division && <Badge>{division.name}</Badge>}
           {conferences &&
             conferences.map((conference) => (
-              <Link
-                href={`/news/${division.slug}/${conference.slug}`}
-                key={conference.slug}
-                className={cn(badgeVariants({ variant: 'default' }), 'font-semibold')}
-              >
-                {conference.shortName ?? conference.name}
-              </Link>
+              <Badge key={conference.shortName}>{conference.shortName}</Badge>
             ))}
         </div>
-
-        <h2 className="text-xl font-semibold tracking-tight">
-          <Link href={`/${slug}`}>{title}</Link>
-        </h2>
-        <div className="flex items-center gap-3 text-sm text-muted-foreground">
-          <div className="flex items-center">
-            <span className="mr-1">By</span>
-            <span className="text-card-foreground">{author.name}</span>
-          </div>
+        <h3 className="mb-2 text-lg font-semibold">
+          <Link href={`/${slug}`} className="transition-colors hover:text-primary" prefetch={false}>
+            {title}
+          </Link>
+        </h3>
+        <div className="flex items-center space-x-2 text-sm text-muted-foreground">
+          <div>{author}</div>
+          <Separator orientation="vertical" className="h-4" />
           <Date dateString={date} />
         </div>
       </div>
     </div>
   )
 }
-
-export default ArticleCard
